build lgpl too, why not?

+++ patches/all_build.sh
@@ -10,17 +10,21 @@ fi
 for dir in sandbox/*/ffmpeg_git*; do
   if [[ -d $dir ]]; then # else there were none, and it passes through the string "sandbox/*..." <sigh>
     cd $dir
+    old_git_version=`git rev-parse HEAD`
     if [[ $1 == "" ]]; then
-      git pull
+      git pull # assume we want them all on git master
     fi 
-    # else don't do git pull as it resets the git hash so forces a rebuild even if it's already previously built to that hash, ex: release build
-    rm -f already*
+    # avoid forcing reconfigure+remake *every time you rerun* this script :|
+    if [[ `git rev-parse HEAD` != $old_git_version ]]; then
+      rm -f already*
+    fi
     cd ../../.. 
   fi
 done
 
 # all are both 32 and 64 bit
 ./cross_compile_ffmpeg.sh --compiler-flavors=multi --disable-nonfree=y --git-get-latest=n --build-ffmpeg-shared=n --build-ffmpeg-static=y $desired_ffmpeg_ver 
+./cross_compile_ffmpeg.sh --compiler-flavors=multi --disable-nonfree=y --git-get-latest=n --build-ffmpeg-shared=y --build-ffmpeg-static=y --enable-gpl=n $desired_ffmpeg_ver 
 ./cross_compile_ffmpeg.sh --compiler-flavors=multi --disable-nonfree=y --git-get-latest=n --build-ffmpeg-shared=y --build-ffmpeg-static=n $desired_ffmpeg_ver
 ./cross_compile_ffmpeg.sh --compiler-flavors=multi --disable-nonfree=y --git-get-latest=n --build-intel-qsv=n --build-ffmpeg-shared=n --build-ffmpeg-static=y $desired_ffmpeg_ver # windows xp :|
 ./cross_compile_ffmpeg.sh --compiler-flavors=multi --disable-nonfree=y --git-get-latest=n --build-ffmpeg-static=y --build-ffmpeg-shared=n --high-bitdepth=y $desired_ffmpeg_ver
